@@ -17,7 +17,7 @@
         <maven.testing.jmh.version>1.21</maven.testing.jmh.version>
         <maven.imc.version>1.0.0-SNAPSHOT</maven.imc.version>
         <maven.jotils.version>0.1.3-SNAPSHOT</maven.jotils.version>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
     <build>
